update service view to handle multiple contacts
